base: clean up and refactor containers.
The containers no longer force the user to use the ptrdiff_t type for
size information. For convenience, size() now always returns an int,
and the size_s() method offers a way to handle arrays with more than
2 billion elements. Internally, we still use ptrdiff_t, though.

Since so much code had to be changed, I took the opportunity to get
rid of capitalised accessors for classes that are lowercase.

+++ src/gpu/renderer.cpp
@@ -234,14 +234,17 @@ void Renderer::AddNew(ivec2 size)
 {
     g_renderers << new Renderer(size);
 }
-ptrdiff_t Renderer::GetCount()
+
+int Renderer::GetCount()
 {
     return g_renderers.count();
 }
-Renderer* Renderer::Get(ptrdiff_t index)
+
+Renderer* Renderer::Get(int index)
 {
     return g_renderers[index];
 }
+
 void Renderer::DestroyAll()
 {
     for (Renderer* renderer : g_renderers)
